+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # Aliases Change Log
 
-
 ## 2.0.1 under development
 
-- no changes in this release.
-
+- New #44: Add method `Aliases::getArray()` that bulk translates path aliases into actual paths (vjik)
 
 ## 2.0.0 April 13, 2021